refactor: 장바구니 내역 영역 파일 분리, components 폴더 이름 정리
