+++ src/refactoring/components/CartCouponSelect.tsx
@@ -1,17 +1,17 @@
 import { Coupon } from "../../types";
 
-interface CouponSelectProps {
+interface CartCouponSelectProps {
   coupons: Coupon[];
   applyCoupon: (coupon: Coupon) => void;
   selectedCoupon: Coupon | null;
 }
 
 // 쿠폰 적용
-export const CouponSelect = ({
+export const CartCouponSelect = ({
   coupons,
   applyCoupon,
   selectedCoupon,
-}: CouponSelectProps) => {
+}: CartCouponSelectProps) => {
   return (
     <div className="mt-6 bg-white p-4 rounded shadow">
       <h2 className="text-2xl font-semibold mb-2">쿠폰 적용</h2>
