+++ src/refactoring/components/ItemDiscountInfo.tsx
@@ -1,0 +1,44 @@
+import { Product } from "../../types";
+
+interface ItemDiscountInfoProps {
+  product: Product;
+  remainingStock: number;
+}
+// 재고, 할인 정보
+export const ItemDiscountInfo = ({
+  product,
+  remainingStock,
+}: ItemDiscountInfoProps) => {
+  const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
+    return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+  };
+
+  return (
+    <>
+      <div className="text-sm text-gray-500 mb-2">
+        <span
+          className={`font-medium ${
+            remainingStock > 0 ? "text-green-600" : "text-red-600"
+          }`}
+        >
+          재고: {remainingStock}개
+        </span>
+        {product.discounts.length > 0 && (
+          <span className="ml-2 font-medium text-blue-600">
+            최대 {(getMaxDiscount(product.discounts) * 100).toFixed(0)}% 할인
+          </span>
+        )}
+      </div>
+      {product.discounts.length > 0 && (
+        <ul className="list-disc list-inside text-sm text-gray-500 mb-2">
+          {product.discounts.map((discount, index) => (
+            <li key={index}>
+              {discount.quantity}개 이상: {(discount.rate * 100).toFixed(0)}%
+              할인
+            </li>
+          ))}
+        </ul>
+      )}
+    </>
+  );
+};
